move firmware parsing into separate crate, use deku, and make no_std compatible
