+++ rust/sdrr-gen/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sdrr-gen"
-version = "0.2.0"
+version = "0.2.1"
 edition = "2024"
 authors = ["Piers Finlayson <piers@piers.rocks>"]
 description = "Software Defined Retro ROM (SDRR) firmware generator"
