+++ rust/sdrr-info/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sdrr-info"
-version = "0.2.0"
+version = "0.2.1"
 edition = "2024"
 authors = ["Piers Finlayson <piers@piers.rocks>"]
 description = "Generates information from Software Defined Retro ROM (SDRR) firmware elf files"
@@ -14,3 +14,4 @@ anyhow = "1.0"
 clap = { version = "4.5", features = ["derive"] }
 chrono = "0.4"
 sdrr-common = { path = "../sdrr-common" }
+sdrr-fw-parser = { path = "../sdrr-fw-parser", features = ["std"] }
